src/Atc.Wpf/ValueConverters/ObjectNotNullToVisibilityVisibleValueConverter.cs
using System;
using System.Globalization;
using System.Windows;
using System.Windows.Data;

namespace Atc.Wpf.ValueConverters
{
    /// <summary>
    /// ValueConverter: Object NotNull To Visibility-Visible.
    /// </summary>
    public class ObjectNotNullToVisibilityVisibleValueConverter : IValueConverter
    {
        /// <inheritdoc />
        public object Convert(object? value, Type targetType, object parameter, CultureInfo culture)
        {
            return value == null
                ? Visibility.Collapsed
                : Visibility.Visible;
        }

        /// <inheritdoc />
        public object ConvertBack(object? value, Type targetType, object parameter, CultureInfo culture)
        {
            return value == null
                ? Visibility.Visible
                : Visibility.Collapsed;
        }
    }
}